Implement 302 See Other redirect
This is necessary to allow one resource to redirect users over to another resource if some criteria is met. I originally tried using a 300 redirect, but the browser aggressively caches this redirect. This, I guess makes sense, and explains why it is good SEO practice...but COME ON! Rails, for its part, uses 301 (Temporary Redirect).